working Search get request
