+++ potluck/src/components/reducers/dataReducer.js
@@ -1,6 +1,4 @@
-import {
-    FETCHING_EVENTS_START, FETCHING_EVENTS_SUCCESS, FETCHING_EVENTS_ERROR,
-    POSTING_LOGIN_START, POSTING_LOGIN_SUCCESS, POSTING_LOGIN_ERROR, POSTING_REGISTRATION_START, POSTING_REGISTRATION_SUCCESS, POSTING_REGISTRATION_ERROR, POSTING_ADDEVENT_START, POSTING_ADDEVENT_SUCCESS, POSTING_ADDEVENT_ERROR, PUTTING_EDITEVENT_START, PUTTING_EDITEVENT_SUCCESS, PUTTING_EDITEVENT_ERROR
+import {POSTING_LOGIN_START, POSTING_LOGIN_SUCCESS, POSTING_LOGIN_ERROR, POSTING_REGISTRATION_START, POSTING_REGISTRATION_SUCCESS, POSTING_REGISTRATION_ERROR, POSTING_ADDEVENT_START, POSTING_ADDEVENT_SUCCESS, POSTING_ADDEVENT_ERROR, PUTTING_EDITEVENT_START, PUTTING_EDITEVENT_SUCCESS, PUTTING_EDITEVENT_ERROR
 } from '../actions/actionsIndex';
 
 export const initialState = {
